pyproject.toml
[tool.poetry]
name = "forum"
version = "0.1.0"
description = ""
authors = ["ndestrieux <n.destrieux@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12"
fastapi = "~0.115"
sqlalchemy = "~2.0"
pydantic-settings = "~2.6"
psycopg2-binary = "~2.9"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
pyjwt = "~2.10"


[tool.poetry.group.dev.dependencies]
black = "*"
isort = "*"
pre-commit = "~4.0"
pytest = "~8.3"
pytest-asyncio = "~0.24"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]
